requirements.txt
# LLM + embeddings
openai>=1.3.0,<1.52.0

# Vector store
# chromadb 0.3.x uses DuckDB+Parquet backend - no SQLite 3.35 requirement
# (0.4+ requires SQLite 3.35 which Python 3.9 on Windows does not ship with)
chromadb==0.3.29
pydantic==1.10.26

# PDF parsing
# pdfminer.six is the actual extractor used in resume_parser.py (handles
# tight-kerned/LaTeX fonts correctly). pdfplumber is kept because it pulls
# in pdfminer as a dependency and may be used for page-level utilities.
pdfminer.six>=20221105
pdfplumber>=0.10.0,<=0.10.4

# Web scraping (JD from URL)
httpx>=0.25.0,<0.28.0
beautifulsoup4>=4.12.0

# UI
streamlit>=1.28.0

# Storage
# sqlmodel 0.0.8 is compatible with pydantic 1.x (required by chromadb 0.3.x)
sqlmodel==0.0.8

# Utilities
python-dotenv>=1.0.0
